requirements.txt
fastapi==0.95.1
PyYAML==6.0
uvicorn==0.22.0
redis==4.5.5
